feat: Add authentication to Source Context fetching via scraping config
