+++ crates/symbolicator-service/src/services/mod.rs
@@ -42,9 +42,9 @@ use self::ppdb_caches::PortablePdbCacheActor;
 use self::sourcemap::SourceMapService;
 use self::symbolication::SymbolicationActor;
 use self::symcaches::SymCacheActor;
+
+pub use self::symbolication::ScrapingConfig;
 pub use fetch_file::fetch_file;
-// TODO: Not super happy about this
-pub use sourcemap_lookup::ScrapingConfig;
 
 pub fn create_service(
     config: &Config,
